Transfer code from helpers to the board-controller
